requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
xgboost==2.0.3
mlflow==2.13.0
shap==0.45.1
imbalanced-learn==0.12.3
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
joblib==1.3.2
evidently==0.7.21
aiofiles==23.2.1
streamlit==1.56.0
apscheduler==3.11.2
pyyaml==6.0.2
